Changed type of R1 because it is only compared to double type in the code

+++ src/MoransI.cpp
@@ -5,7 +5,7 @@ using namespace Rcpp;
 double MoransI(SEXP mat1, SEXP r1) {
 
   NumericMatrix Mat1(mat1);
-  int R1 = as<int>(r1);
+  double R1 = as<double>(r1);
   int Rows = Mat1.nrow();
   int Cols = Mat1.ncol();
   int n1 = Rows * Cols;
